meson.build.o7QLX02 Attribute changed meson.build.o7QLX02 src/Application.vala.oueOPYm Attribute changed src/Application.vala.oueOPYm src/Builder4/About.bjs.ohhfP9G Attribute changed src/Builder4/About.bjs.ohhfP9G src/Builder4/About.vala.oQqbnk1 Attribute changed src/Builder4/About.vala.oQqbnk1 src/Builder4/DialogFiles.bjs.olqtmul Attribute changed src/Builder4/DialogFiles.bjs.olqtmul src/Builder4/DialogFiles.vala.o4yEZGF Attribute changed src/Builder4/DialogFiles.vala.o4yEZGF src/Builder4/Editor.bjs.o73T50Z Attribute changed src/Builder4/Editor.bjs.o73T50Z src/Lsp.vala.oBBEcSp Attribute changed src/Lsp.vala.oBBEcSp src/Makefile.am.oWZHMmJ Attribute changed src/Makefile.am.oWZHMmJ src/Builder4/Editor.vala.oU2K3jk Attribute changed src/Builder4/Editor.vala.oU2K3jk src/Builder4/GtkView.bjs.oHdUFEE Attribute changed src/Builder4/GtkView.bjs.oHdUFEE src/Builder4/GtkView.vala.oemSKTY Attribute changed src/Builder4/GtkView.vala.oemSKTY src/Builder4/MainWindow.bjs.oRluDaj Attribute changed src/Builder4/MainWindow.bjs.oRluDaj src/Builder4/MainWindow.vala.o6abMqD Attribute changed src/Builder4/MainWindow.vala.o6abMqD src/Builder4/ValaProjectSettingsPopover.bjs.olg2TVX Attribute changed src/Builder4/ValaProjectSettingsPopover.bjs.olg2TVX src/Builder4/ValaProjectSettingsPopover.vala.oSEaYpi Attribute changed src/Builder4/ValaProjectSettingsPopover.vala.oSEaYpi src/Builder4/WindowLeftProps.bjs.o7LyQMC Attribute changed src/Builder4/WindowLeftProps.bjs.o7LyQMC src/Builder4/WindowLeftProps.vala.osUPTcX Attribute changed src/Builder4/WindowLeftProps.vala.osUPTcX src/Builder4/WindowLeftTree.bjs.ozROCGg Attribute changed src/Builder4/WindowLeftTree.bjs.ozROCGg src/Builder4/WindowLeftTree.vala.o6TNZ9z Attribute changed src/Builder4/WindowLeftTree.vala.o6TNZ9z src/Builder4/WindowRooView.bjs.obkwkvT Attribute changed src/Builder4/WindowRooView.bjs.obkwkvT src/Builder4/WindowRooView.vala.oupErTc Attribute changed src/Builder4/WindowRooView.vala.oupErTc src/Builder4/WindowState.vala.opE7lww Attribute changed src/Builder4/WindowState.vala.opE7lww src/Builder4/About.bjs src/Builder4/About.vala src/Builder4/DialogFiles.bjs src/Builder4/DialogFiles.vala src/Builder4/Editor.bjs src/Builder4/Editor.vala src/Builder4/GtkView.bjs src/Builder4/GtkView.vala src/Builder4/MainWindow.bjs src/Builder4/MainWindow.vala src/Builder4/ValaProjectSettingsPopover.bjs src/Builder4/ValaProjectSettingsPopover.vala src/Builder4/WindowLeftProps.bjs src/Builder4/WindowLeftProps.vala src/Builder4/WindowLeftTree.bjs src/Builder4/WindowLeftTree.vala src/Builder4/WindowRooView.bjs src/Builder4/WindowRooView.vala src/Builder4/WindowState.vala src/JsRender/Gtk.vala.o2n1F8P Attribute changed src/JsRender/Gtk.vala.o2n1F8P src/JsRender/JsRender.vala.o7fcEK9 Attribute changed src/JsRender/JsRender.vala.o7fcEK9 src/JsRender/Node.vala.omjE5lt Attribute changed src/JsRender/Node.vala.omjE5lt src/JsRender/NodeToGlade.vala.ov3LSRM Attribute changed src/JsRender/NodeToGlade.vala.ov3LSRM src/JsRender/NodeToVala.vala.ogbyln6 Attribute changed src/JsRender/NodeToVala.vala.ogbyln6 src/JsRender/Gtk.vala src/JsRender/JsRender.vala src/JsRender/Node.vala src/JsRender/NodeToGlade.vala src/JsRender/NodeToVala.vala src/Palete/CompileError.vala.oTJC1U2 Attribute changed src/Palete/CompileError.vala.oTJC1U2 src/Palete/CompletionProvider.vala.ow1Putm Attribute changed src/Palete/CompletionProvider.vala.ow1Putm src/Palete/LanguageClient.vala.ovgsk1F Attribute changed src/Palete/LanguageClient.vala.ovgsk1F src/Palete/LanguageClientDummy.vala.o2EYIxZ Attribute changed src/Palete/LanguageClientDummy.vala.o2EYIxZ src/Palete/LanguageClientVala.vala.oJkVp4i Attribute changed src/Palete/LanguageClientVala.vala.oJkVp4i src/Palete/ValaCompileRequest.vala.os3UTNC Attribute changed src/Palete/ValaCompileRequest.vala.os3UTNC src/Palete/ValaSource.vala.oND2UwW Attribute changed src/Palete/ValaSource.vala.oND2UwW src/Palete/CompileError.vala src/Palete/CompletionProvider.vala src/Palete/LanguageClient.vala src/Palete/LanguageClientDummy.vala src/Palete/LanguageClientVala.vala src/Palete/ValaCompileRequest.vala src/Palete/ValaSource.vala meson.build src/Spawn.vala.oVndLod Attribute changed src/Spawn.vala.oVndLod src/Application.vala src/Lsp.vala src/Makefile.am src/Spawn.vala src/codegen/valaccodecompiler.vala.o2utUax Attribute changed src/codegen/valaccodecompiler.vala.o2utUax src/codegen/valaccodecompiler.vala src/Project/Gtk.vala.o4D83fg Attribute changed src/Project/Gtk.vala.o4D83fg src/Project/GtkValaSettings.vala.oBuRPYz Attribute changed src/Project/GtkValaSettings.vala.oBuRPYz src/Project/Project.vala.o48SLGT Attribute changed src/Project/Project.vala.o48SLGT src/Project/Gtk.vala src/Project/GtkValaSettings.vala src/Project/Project.vala

+++ src/Palete/ValaCompileRequest.vala
@@ -165,7 +165,9 @@ namespace Palete {
 		}
 		
 		public void cancel() {
-			Posix.kill(this.compiler.pid, 9);
+			if (this.compiler != null && this.compiler.pid > 0) {
+				Posix.kill(this.compiler.pid, 9);
+			}
 			this.compiler = null;
 			this.deleteTemp();
 		}
@@ -196,10 +198,16 @@ namespace Palete {
 			this.compiler.isZombie();
 			GLib.debug("compile got %s", output);
 			if (output == "") {
-			 	BuilderApplication.showSpinner(false);
+    	        BuilderApplication.showSpinner("face-sad", "compile failed - no error message?");
 			 	return;
 		 	}
-		 	
+		 	if (this.requestType == ValaCompileRequestType.RUN) {
+    	        BuilderApplication.showSpinner("");
+				this.execResult();
+				return;
+			}
+			
+			// below is not used anymore - as we dont use this
 			try { 
 				//GLib.debug("GOT output %s", output);
 				
@@ -208,12 +216,12 @@ namespace Palete {
 				var node = pa.get_root();
 
 				if (node.get_node_type () != Json.NodeType.OBJECT) {
-					BuilderApplication.showSpinner(false);
+					BuilderApplication.showSpinner("");
 					return;
 				}
 				var ret = node.get_object ();	
 				//CompileError.parseCompileResults(this,ret);
-				BuilderApplication.showSpinner(false);
+					BuilderApplication.showSpinner("");
 				
 			
 				
@@ -221,12 +229,9 @@ namespace Palete {
 				
 			} catch (GLib.Error e) {
 				GLib.debug("parsing output got error %s", e.message);
-				BuilderApplication.showSpinner(false);
+				BuilderApplication.showSpinner("");
 				return;
 				
-			}
-			if (this.requestType == ValaCompileRequestType.RUN) {
-				this.execResult();
 			}
 		}
 		 
@@ -269,10 +274,9 @@ namespace Palete {
 			var contents = this.alt_code == "" ? this.file.toSourceCode() : this.generateTempContents();
 			
 		 	Javascript.singleton().validate(contents, this.file );
-			
-		 	
 			 
-			BuilderApplication.showSpinner(false);
+			 
+			BuilderApplication.showSpinner("");
 			BuilderApplication.updateCompileResults();
 			
 			//this.queue.onCompileComplete(this);
@@ -289,7 +293,7 @@ namespace Palete {
 			var cn = "/proc/%d/task/%d/children".printf(pid,pid);
 			if (!FileUtils.test(cn, GLib.FileTest.EXISTS)) {
 				GLib.debug("%s doesnt exist - killing %d", cn, pid);
-				Posix.kill(pid, 9);
+				 Posix.kill(pid, 9);
 				return;
 			}
 			string cpids = "";
@@ -305,10 +309,10 @@ namespace Palete {
 				// skip
 			}
 			GLib.debug("killing %d", pid);	
-			Posix.kill(pid, 9);
+			//Posix.kill(pid, 9);
 		}
 		
-		int terminal_pid = 0;
+		public int terminal_pid = 0;
 		public void execResult()
 		{
 			  	
@@ -319,22 +323,22 @@ namespace Palete {
 			var pr = (Project.Gtk) this.file.project;
 			var cg =  pr.compilegroups.get(exe);
 			
-			if (!GLib.FileUtils.test(exe, GLib.FileTest.EXISTS)) {
-				print("Missing output file: %s\n",exe);
+			var exbin = pr.path + "/build/" + exe;
+			if (!GLib.FileUtils.test(exbin, GLib.FileTest.EXISTS)) {
+				GLib.debug("Missing output file: %s\n",exbin);
 				return;
 			}
-			var gdb_cfg= pr.path + "/build/.gdb-script";
+			var gdb_cfg = pr.path + "/build/.gdb-script";
 			if (!GLib.FileUtils.test(gdb_cfg, GLib.FileTest.EXISTS)) {
 				pr.writeFile("build/.gdb-script", "set debuginfod enabled off\nr");
 			}
-			
-			
+			 
 			
 			string[] args = "/usr/bin/gnome-terminal --disable-factory --wait -- /usr/bin/gdb -x".split(" ");
 
 			args+= gdb_cfg;
  
-			args += exe;
+			args += exbin;
 			if (cg.execute_args.length > 0) {
    			 	args+= "--args";
 				var aa = cg.execute_args.split(" ");
